아이디를 admin으로 가입시 안되게 수정

+++ bank/src/net/wkbae/assignment/wp/DoSignServlet.java
@@ -21,7 +21,7 @@ public class DoSignServlet extends HttpServlet {
 		String accountnum = request.getParameter("accountnum");
 		
 		File f = new File("c:/bankuser/" + id + ".txt");
-		if(!f.isFile()) {
+		if(!f.isFile() && !id.equals("admin")) {
 			try(FileOutputStream fos = new FileOutputStream(f);
 			    DataOutputStream dos = new DataOutputStream(fos)) {
 				dos.writeUTF(name);
